pyproject.toml
[project]
name = "multi-agent-data-collector"
version = "0.1.0"
description = "Project to create a multi-agent data collector the idea is that this data collector will be able to go to the internet and search and find results and fill data into some sort of database this is to gain a better understanding of how multi-agent systems work "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.11.0",
    "crewai[tools]>=0.108.0",
    "firecrawl-py>=1.15.0",
    "loguru>=0.7.3",
    "numpy>=2.2.4",
    "openai>=1.70.0",
    "pandas>=2.2.3",
    "pydantic>=2.0.0",
    "requests>=2.32.3",
]
